Clarify Kafka message format expectations in comments
Added comments to specify that Kafka message keys and values should be strings. Included details on the consumer's deserialization process and the message format used in the Kafka producer, which is set to `<string, string>` with "VehicleId" as the key and a JSON representation of the `VehicleLocation` object as the value.